requirements.txt
flask==3.0.0
twilio==8.10.0
google-generativeai==0.3.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
gspread==5.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0